+++ src/darkgraylib/command_line.py
@@ -1,15 +1,19 @@
 """Command line parsing for the ``darker`` and ``graylint`` binaries."""
 
+from __future__ import annotations
+
 import sys
 from argparse import SUPPRESS, ArgumentParser, Namespace
 from functools import partial
-from typing import Any, List, Optional, Tuple, Type, TypeVar, Protocol
+from typing import Any, Callable, Protocol, TypeVar
 
 from darkgraylib import help as hlp
 from darkgraylib.argparse_helpers import (
     LogLevelAction,
     NewlinePreservingFormatter,
-    OptionsForReadmeAction, UpdateReadmeAction, VerifyReadmeAction,
+    OptionsForReadmeAction,
+    UpdateReadmeAction,
+    VerifyReadmeAction,
 )
 from darkgraylib.config import (
     BaseConfig,
@@ -83,7 +87,10 @@ def make_argument_parser(
 
 
 def add_parser_argument(
-    parser: ArgumentParser, help_text: Optional[str], *name_or_flags: str, **kwargs: Any
+    parser: ArgumentParser,
+    help_text: str | None,
+    *name_or_flags: str,
+    **kwargs: Any,  # noqa: ANN401
 ) -> None:
     """Add an argument to the parser
 
@@ -110,10 +117,11 @@ class ArgumentParserFactory(Protocol):  # pylint: disable=too-few-public-methods
 
 def parse_command_line(
     argument_parser_factory: ArgumentParserFactory,
-    argv: Optional[List[str]],
+    argv: list[str] | None,
     section_name: str,
-    config_type: Type[T],
-) -> Tuple[Namespace, T, T]:
+    config_type: type[T],
+    load_config_hook: Callable[[T], None] | None = None,
+) -> tuple[Namespace, T, T]:
     """Return the parsed command line, using defaults from a configuration file
 
     Also return the effective configuration which combines defaults, the configuration
@@ -129,6 +137,9 @@ def parse_command_line(
     :param config_type: The type of the configuration object to be returned. For Darker,
                         this should be ``DarkerConfig``, for Graylint
                         ``GraylintConfig``.
+    :param load_config_hook: A hook to call after loading the configuration file. This
+                             is used to warn about configuration options which are
+                             deprecated in the configuration file.
     :return: A tuple of the parsed command line, the effective configuration, and the
              set of modified configuration options from the defaults.
 
@@ -146,6 +157,8 @@ def parse_command_line(
     #    directory if no paths were given. Load Darker or Graylint configuration from
     #    it.
     pyproject_config = load_config(args.config, args.src, section_name, config_type)
+    if load_config_hook:
+        load_config_hook(pyproject_config)
 
     # 3. The PY_COLORS, NO_COLOR and FORCE_COLOR environment variables override the
     #    `--color` command line option.